+env var for API URL in fetch, +https support, parcel is req
add: https server for production.
update: apiFetch to use environment variables to api url.
add: parcel is now a dependency as the cli is needed to build the bundle
adding the url prior to starting the server. will find a way to remove
this dependency later.
dependencies: parcel, it added browserfiyx2 and process.