+update my account view, using updateMe, updateMyPassword, and deleteMe
Learned a lot during this run, incorporated the form attribute to
buttons on all the views and is now used to get the formData. Having
multiple forms on one page showed the importance of this.

add: myAccount view with functionality to update the user info, update
password and delete account.
update: views with form attribute and button tags.
add: nav-bar navigation, which is used in myAccount view.